apps/ai-backend/requirements.txt
fastapi
uvicorn
pydantic
motor
beanie
python-dotenv
langchain-ollama
langchain-core
langgraph
requests
dnspython
